Revert "msm: vidc: Increase extradata size to 16k"
A merged commit by CodeAurora before the kernel source was released (as older commits were missing) caused video recording to not work on older MSM8916 devices such as the Galaxy Grand Prime.

Using Galaxy-MSM8916's kernel source as a point-of-reference, we will try this and see if video recording works and/or stabilizes now.